+++ src/App/Http/Middleware/Authenticate.php
@@ -46,7 +46,23 @@ class Authenticate
         
         // Check if user has permission
         if ($user != null) {
-            if ($user->hasAccess($request->path())) {
+            if (! $user->activated) {
+                // User logged in but was deactivated after
+                // Log out this user and bring to login page
+                Auth::logout();
+                return redirect()->guest('login');
+            }
+            // Check the type of request
+            $type = 'view';
+            if ($request->is('admin/*/create')) {
+                $type = 'create';
+            } elseif ($request->is('admin/*/edit/*')) {
+                $type = 'update';
+            } elseif ($request->is('admin/*/delete/*')) {
+                $type = 'delete';
+            }
+            // Proceed to check user permission
+            if ($user->hasAccess($request->path(), $type)) {
                 // Save login time
                 $user->last_login = date('Y-m-d H:i:s');
                 $user->save();